requirements.txt
fastapi
uvicorn[standard]
snowflake-connector-python
python-multipart
